Moved MailStep to basic-steps
